[#50] Remove dump from schema
